backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["victorbost"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "5.0.0"
djangorestframework = "3.14.0"
django-cors-headers = "4.3.1"
django-filter = "23.5"
djangorestframework-simplejwt = "5.5.1"
psycopg2-binary = "2.9.9"
python-dotenv = "1.0.0"
pillow = "10.2.0"
boto3 = "1.34.17"
django-storages = "1.14.2"
requests = "2.31.0"
gunicorn = "21.2.0"
whitenoise = "6.6.0"
sentry-sdk = "1.39.2"
djangorestframework-gis = "^1.2.0"
drf-yasg = "^1.21.11"


[tool.poetry.group.dev.dependencies]
black = "23.12.1"
flake8 = "7.0.0"
pytest = "7.4.4"
pytest-django = "4.7.0"
pytest-cov = "4.1.0"
factory-boy = "3.3.0"
faker = "22.0.0"
ipython = "8.19.0"
django-debug-toolbar = "4.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
preview = true
line-length = 79
enable-unstable-feature = ["string_processing"]
target-version = ["py310", "py311", "py312", "py313"]
